[Client] Extract common PriceField component
